Adjust the time used for domestic QR code generation with clock skew
For various reasons the device clock might be set incorrectly, whilst verification of the QR code by the scanner app requires an accurate timestamp for the challenge. To make generation of the code more resilient, the server time is used to adjust the timestamp used for generation to the "actual" timestamp.

As this timestamp is determined when the configuration is retrieved and not stored, the adjustment currently requires a network connection.