Refactor DialogUtil / DialogFragment
Remove the DialogFragmentViewModel and pass arguments directly to the fragment.
Previously calling the view model instance was available before the DialogFragment was actually created, due to how Koin instantiated the ViewModel. Now that it is using the androidx viewModels() delegate, this is no longer the case.

In stead, the arguments are passed as fragment arguments and the result is handled using the result API.
State is (unfortunately) still lost when the activity is recreated, because we can't keep a reference to the callbacks passed into DialogUtil. Fixing that would require a more standard approach for launching and handling the dialog result.